+++ app.js
@@ -784,6 +784,7 @@ import { renderers as RV1413 } from './views/group-v1413.js';
 import { renderers as RV1414 } from './views/group-v1414.js';
 import { renderers as RV1416 } from './views/group-v1416.js';
 import { renderers as RV1417 } from './views/group-v1417.js';
+import { renderers as RV1418 } from './views/group-v1418.js';
 import { renderers as RV1396 } from './views/group-v1396.js';
 import { renderers as RV164 } from './views/group-v164.js';
 import { renderers as RV165 } from './views/group-v165.js';
@@ -872,7 +873,7 @@ const RENDERERS = { ...RA, ...RB, ...RC, ...RE, ...RF, ...RG, ...RH, ...RI, ...R
   ...RV902,
   ...RV903,
   ...RV905,
-  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV63, ...RPALINT };
+  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV63, ...RPALINT };
 
 // ----- Utility registry ----------------------------------------------------
 // Source of truth for routes, names, group, audiences, and clinical flag.
@@ -1773,6 +1774,7 @@ const UTILITIES = [
   { id: 'wiltse-spondylolisthesis', name: 'Wiltse Classification (Spondylolisthesis)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'russell-taylor-subtroch',  name: 'Russell-Taylor Classification (Subtrochanteric Fracture)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'vancouver-periprosthetic', name: 'Vancouver Classification (Periprosthetic Femoral Fracture)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
+  { id: 'paprosky-acetabular', name: 'Paprosky Classification (Acetabular Bone Loss)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'barrack-cement',           name: 'Barrack Grade (Femoral Cement Mantle)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'dejour-trochlea',          name: 'Dejour Classification (Trochlear Dysplasia)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'samilson-prieto',          name: 'Samilson-Prieto Grade (Shoulder Dislocation Arthropathy)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },